fix(align): use u64 for band area to prevent overflow on 32-bit WASM
Band area calculation accumulates stripe widths across all reference positions (R × W). For pathological inputs like concatenated genomes (e.g., 179kb query vs 30kb reference), this yields ~3.7 billion, which exceeds `u32::MAX` (~4.3B) and causes overflow on 32-bit WASM.

Previous fix (9d95d0ac5) used `saturating_add()` to prevent panic, but saturated values produce misleading error messages. Using `u64` ensures accurate calculation on both 32-bit (WASM) and 64-bit platforms without saturation artifacts.

- Change `max_band_area` parameter from `usize` to `u64`
- Change `create_alignment_band()` return type from `(Vec<Stripe>, usize)` to `(Vec<Stripe>, u64)`
- Accumulate band area as `u64` in `regularize_stripes()`
- Update JSON schemas to reflect `uint64` format

Ref: https://github.com/nextstrain/nextclade/issues/1749

+++ packages/nextclade/src/align/params.rs
@@ -77,7 +77,7 @@ pub struct AlignPairwiseParams {
 
   /// Maximum area of the band in the alignment matrix. Alignments with large bands are slow to compute and require substantial memory. Alignment of sequences requiring bands with area larger than this value, will not be attempted and a warning will be emitted.
   #[clap(long)]
-  pub max_band_area: usize,
+  pub max_band_area: u64,
 
   /// Retry seed matching step with a reverse complement if the first attempt failed
   #[clap(long)]
